extract verity_* function to ics02/client_def.rs
