+++ modules/src/clients/ics07_tendermint/client_state.rs
@@ -188,20 +188,6 @@ impl ClientState {
 
         Ok(())
     }
-
-    /// Verify that the client is at a sufficient height and unfrozen at the given height
-    pub fn verify_height(&self, height: Height) -> Result<(), Error> {
-        if self.latest_height < height {
-            return Err(Error::insufficient_height(self.latest_height(), height));
-        }
-
-        match self.frozen_height {
-            Some(frozen_height) if frozen_height <= height => {
-                Err(Error::client_frozen(frozen_height, height))
-            }
-            _ => Ok(()),
-        }
-    }
 }
 
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
@@ -218,6 +204,10 @@ impl crate::core::ics02_client::client_state::ClientState for ClientState {
 
     fn client_type(&self) -> ClientType {
         ClientType::Tendermint
+    }
+
+    fn proof_specs(&self) -> &ProofSpecs {
+        &self.proof_specs
     }
 
     fn latest_height(&self) -> Height {
